Moved mixins in a package and added add_fluents and add_actions methods
