+++ unified_planning/model/mixins/__init__.py
@@ -1,0 +1,26 @@
+# Copyright 2022 AIPlan4EU project
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+from unified_planning.model.mixins.actions_set import ActionsSetMixin
+from unified_planning.model.mixins.fluents_set import FluentsSetMixin
+from unified_planning.model.mixins.objects_set import ObjectsSetMixin
+from unified_planning.model.mixins.user_types_set import UserTypesSetMixin
+
+__all__ = [
+    'ActionsSetMixin',
+    'FluentsSetMixin',
+    'ObjectsSetMixin',
+    'UserTypesSetMixin'
+]
